+++ WebAPICalc4/Utilities/Calculator.cs
@@ -8,35 +8,35 @@ namespace WebAPICalc4.Utilities
 {
     public class Calculator : ICalculator
     {
-        public int Add(int a, int b)
+        public long Add(int a, int b)
         {
-            return (a + b);
+            return (long)a + (long)b;
         }
 
-        public int Sub(int a, int b)
+        public long Sub(int a, int b)
         {
-            return (a - b);
+            return (long)a - (long)b;
         }
 
-        public int Mult(int a, int b)
+        public long Mult(int a, int b)
         {
-            return (a * b);
+            return (long)a * (long)b;
         }
 
-        public int Div(int a, int b)
+        public long Div(int a, int b)
         {
             if (b == 0)
-                return Int32.MaxValue;
-            return (a / b);
+                return long.MaxValue;
+            return (long)a / (long)b;
         }
 
-        public int Eval(string str)
+        public long Eval(string str)
         {
             int i;
 
             i = 0;
             str = str.Replace(" ", "");
-            return (new FunctionsForStringCalc().parse_sum(str, ref i));
+            return (long)(new FunctionsForStringCalc().parse_sum(str, ref i));
         }
     }
 }
